Update sample client and server with new changes.

+++ server.lua
@@ -1,30 +1,67 @@
 #!/usr/local/bin/lua
 
-net   = require "org.conman.net"
-fsys  = require "org.conman.fsys"
-unix  = require "org.conman.unix"
-errno = require "org.conman.errno"
-sig   = require "org.conman.process".sig
-        require "readacl"
+syslog = require "org.conman.syslog"
+net    = require "org.conman.net"
+fsys   = require "org.conman.fsys"
+unix   = require "org.conman.unix"
+errno  = require "org.conman.errno"
+proc   = require "org.conman.process"
+sig    = require "org.conman.process".sig
+         require "readacl"
 
-sig.catch(sig.INT)
-fsys.umask("--x--x--x")
+syslog.open('IPport','daemon')
+syslog('notice',"IPport starting up")
+local g_null = fsys.open("/dev/null","rw")
+local g_sock
 
-os.remove("/tmp/acl")
-laddr = net.address("/tmp/acl")
-sock  = net.socket(laddr.family,'udp')
-sock:bind(laddr)
+-- *********************************************************************
 
-recvcred(sock)
+function log_success(remote,cred,addr,stype)
+  syslog('notice',string.format("success: rem=%s uid=%s gid=%s family=%s addr=%s port=%s type=%s",
+  		tostring(remote),
+      		cred.uid,
+      		cred.gid,
+      		addr.family,
+      		addr.addr,
+      		addr.port,
+      		stype))      		
+end
 
-while true do
-  remote,data,cred,err = readcred(sock)
+-- *********************************************************************
+
+function log_failure(remote,cred,addr,stype,err)
+  syslog('err',string.format("failure: err=%q rem=%s uid=%s gid=%s family=%s addr=%s port=%s type=%s",
+  		errno.strerror(err),
+  		tostring(remote),
+      		cred.uid,
+      		cred.gid,
+      		addr.family,
+      		addr.addr,
+      		addr.port,
+      		stype))
+end
+
+-- *********************************************************************
+
+function main()
+  local remote
+  local data
+  local cred
+  local err
+  local addr
+  local stype
+  local sock
+  local _
   
+  remote,data,cred,err = readcred(g_sock)
   if err ~= 0 then
-    print("readcred()",errno.strerror(err))    
-    sock:close()
-    os.remove("/tmp/acl")
-    os.exit(1)
+    if err == errno.EINTR then
+      g_sock:close()
+      os.remove("/tmp/acl")
+      os.exit(1)
+    end
+    syslog('crit',string.format("readcred() = %s",errno.strerror(err)))
+    return main()
   end
   
   if unix.users[cred.uid] ~= nil then
@@ -35,12 +72,42 @@ while true do
     cred.gid = unix.groups[cred.gid].name
   end
   
-  print(string.format("cred remote=%q pid=%d uid=%s gid=%s",
-  	remote.addr,
-  	cred.pid,
-  	cred.uid,
-  	cred.gid
-  ))
+  addr,stype,err = acl_decode(data)
+  if err ~= 0 then
+    log_failure(remote,cred,addr,stype,err)
+    sendfd(g_sock,remote,g_null:fd(),err)
+    return main()
+  end
   
+  sock,err = net.socket(addr.family,stype)
+  if err ~= 0 then
+    log_failure(remote,cred,addr,stype,err)
+    sendfd(g_sock,remote,g_null:fd(),err)
+    return main()
+  end
+  
+  err = sock:bind(addr)
+  if err ~= 0 then
+    log_failure(remote,cred,addr,stype,err)
+    sendfd(g_sock,remote,g_null:fd(),err)
+    sock:close()
+    return main()
+  end
+
+  log_success(remote,cred,addr,stype)  
+  sendfd(g_sock,remote,sock:fd())
+  sock:close()
+  return main()
 end
 
+-- *************************************************************************  
+
+sig.catch(sig.INT)
+fsys.umask("--x--x--x")
+os.remove("/tmp/acl")
+laddr = net.address("/tmp/acl")
+g_sock  = net.socket(laddr.family,'udp')
+g_sock:bind(laddr)
+recvcred(g_sock)
+
+main()
